+++ packages/json-rpc/ws-provider/src/node.ts
@@ -1,12 +1,15 @@
-// @ts-ignore
 import { WebSocket } from "ws"
 import { getInternalWsProvider } from "./ws-provider"
 export type { JsonRpcProvider } from "@polkadot-api/json-rpc-provider"
 
+export * from "./types"
+
 class WS extends WebSocket {
   close() {
-    ;(this as any).terminate()
+    this.terminate()
   }
 }
 
-export const getWsProvider = getInternalWsProvider(WS as any)
+export const getWsProvider = getInternalWsProvider(
+  WS as unknown as typeof globalThis.WebSocket,
+)