+++ packages/json-rpc/ws-provider/src/ws-provider.ts
@@ -1,35 +1,5 @@
-import type { JsonRpcProvider } from "@polkadot-api/json-rpc-provider"
 import { getSyncProvider } from "@polkadot-api/json-rpc-provider-proxy"
-
-export enum WsEvent {
-  CONNECTING,
-  CONNECTED,
-  ERROR,
-  CLOSE,
-}
-export type WsConnecting = {
-  type: WsEvent.CONNECTING
-  uri: string
-  protocols?: string | string[]
-}
-export type WsConnected = {
-  type: WsEvent.CONNECTED
-  uri: string
-  protocols?: string | string[]
-}
-export type WsError = {
-  type: WsEvent.ERROR
-  event: any
-}
-export type WsClose = {
-  type: WsEvent.CLOSE
-  event: any
-}
-export type StatusChange = WsConnecting | WsConnected | WsError | WsClose
-export type WsJsonRpcProvider = JsonRpcProvider & {
-  switch: (uri?: string, protocol?: string[]) => void
-  getStatus: () => StatusChange
-}
+import { StatusChange, WsJsonRpcProvider, WsEvent } from "./types"
 
 export interface GetWsProviderInput {
   (
@@ -53,6 +23,7 @@ const timeoutError: StatusChange = {
 }
 
 const noop = () => {}
+
 export const getInternalWsProvider = (
   WebsocketClass: typeof WebSocket,
 ): GetWsProviderInput => {