+++ packages/client/CHANGELOG.md
@@ -11,6 +11,9 @@
 - Avoid throwing when creating a transaction that has an unknown signed-extension.
 - Validate transactions on Runtimes that only support metadata V14.
 
+- **WS Provider:**
+  - Export missing types
+
 ## 1.4.1 - 2024-10-05
 
 ### Fixed